LPS-61952 Move non-generated files to new package
