+++ portal-service/src/com/liferay/announcements/kernel/model/packageinfo
