+++ portal-service/src/com/liferay/announcements/kernel/packageinfo
