+++ portal-service/src/com/liferay/announcements/kernel/service/packageinfo
