+++ portal-service/src/com/liferay/announcements/kernel/util/packageinfo
